+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './batch-submitter'
 export * from './utils'
-export * from './transaciton-chain-contract'
+export * from './transaction-chain-contract'
 export * from './types'
